Incomplete refactor checkpoint TODO: clean up Tests function to include updated TestConfigs
